Restructure sandboxes module: lift session lifecycle, split files
Six code-quality moves per the thermo-nuclear review:

1. Split base_sandbox.py (751 LOC) into focused modules — process.py
   (SandboxProcess, SandboxOutput, SandboxExecError), snapshot.py
   (BaseSandboxSnapshot), session.py (SandboxSession with lifecycle),
   base.py (BaseSandbox, Config, Settings, Flavor). Public surface
   stays stable via __init__.py re-exports.

2. Lift session lifecycle onto SandboxSession base — __init__ takes
   id/parent/forward_logs, __enter__/__exit__ are concrete, _log_ctx
   state owned in one place, _close_log_ctx() helper for subclasses.
   LocalSandboxSession drops ~25 lines of duplicated plumbing.

3. Move step-metadata logging to SandboxSession._on_enter with a
   _dashboard_url() -> Optional[str] hook for flavors. Generic fields
   (sandbox_session_id, sandbox_flavor, Uri-typed dashboard_url) land
   for free in every flavor.

4. New BaseSandbox._resolve_forward_logs() one-shot helper so
   create_session/attach/restore collapse to a single call each.

5. New BaseSandbox.effective_settings(override) helper that handles
   the generic 'override-or-pull-step-context -> layer config defaults
   -> coerce to settings_class' pipeline. LocalSandbox's bespoke
   _effective_settings is gone.

6. Narrowed pull_step_settings exception scope to ValidationError,
   grouped BaseSandbox methods by responsibility, made SandboxSession.id
   an init-set instance attribute (no more class-attr-as-annotation
   masquerade).

73 unit tests pass; ruff/pydoclint/mypy clean on src/zenml/sandboxes/.
